+++ scopes/pkg/pkg/index.ts
@@ -1,3 +1,4 @@
 export type { PkgMain, PackageJsonProps } from './pkg.main.runtime';
-export { PackageDependency, PackageDependencyFactory } from './package-dependency';
-export { PkgAspect } from './pkg.aspect';
+export type { PackageDependency, PackageDependencyFactory } from './package-dependency';
+export { PkgAspect as default, PkgAspect } from './pkg.aspect';
+export { PkgUI } from './pkg.ui.runtime';
