+++ scopes/pkg/pkg/pkg.ui.runtime.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { UIRuntime } from '@teambit/ui';
+import { ConsumePlugin } from '@teambit/component/ui/menu';
+import { Install } from '@teambit/ui-foundation.ui.use-box.menu';
+import ComponentAspect, { ComponentUI } from '@teambit/component';
+import { PkgAspect } from './pkg.aspect';
+
+export class PkgUI {
+  static runtime = UIRuntime;
+
+  static dependencies = [ComponentAspect];
+
+  static async provider([componentUI]: [ComponentUI]) {
+    const npm = new PkgUI();
+    componentUI.registerConsumeMethod(npm.npmMethod);
+    return npm;
+  }
+
+  private npmMethod: ConsumePlugin = (comp) => {
+    const registry = comp.packageName.split('/')[0];
+    const packageVersion = comp.version === comp.latest ? '' : `@${comp.version}`;
+    return {
+      Title: <img style={{ width: '30px' }} src="http://static.bit.dev/brands/logo-npm-new.svg" />,
+      Component: (
+        <Install
+          config={`npm config set '${registry}:registry' https://node.bit.dev`}
+          componentName={comp.id.name}
+          packageManager="npm"
+          copyString={`npm i ${comp.packageName}${packageVersion}`}
+          registryName={registry}
+        />
+      ),
+      order: 10,
+    };
+  };
+}
+
+export default PkgUI;
+
+PkgAspect.addRuntime(PkgUI);
